+++ platformio/managers/core.py
@@ -24,7 +24,7 @@ from platformio.proc import get_pythonexe_path
 from platformio.project.config import ProjectConfig
 
 CORE_PACKAGES = {
-    "contrib-piohome": "~3.2.0",
+    "contrib-piohome": "~3.2.1",
     "contrib-pysite": "~2.%d%d.0" % (sys.version_info.major, sys.version_info.minor),
     "tool-unity": "~1.20500.0",
     "tool-scons": "~2.20501.7" if PY2 else "~3.30102.0",
@@ -100,14 +100,26 @@ def update_core_packages(only_check=False, silent=False):
     return True
 
 
-def inject_contrib_pysite():
-    from site import addsitedir  # pylint: disable=import-outside-toplevel
+def inject_contrib_pysite(verify_openssl=False):
+    # pylint: disable=import-outside-toplevel
+    from site import addsitedir
 
     contrib_pysite_dir = get_core_package_dir("contrib-pysite")
     if contrib_pysite_dir in sys.path:
-        return
+        return True
     addsitedir(contrib_pysite_dir)
     sys.path.insert(0, contrib_pysite_dir)
+
+    if not verify_openssl:
+        return True
+
+    try:
+        # pylint: disable=import-error,unused-import,unused-variable
+        from OpenSSL import SSL
+    except:  # pylint: disable=bare-except
+        build_contrib_pysite_deps(get_core_package_dir("contrib-pysite"))
+
+    return True
 
 
 def build_contrib_pysite_deps(target_dir):
@@ -126,13 +138,13 @@ def build_contrib_pysite_deps(target_dir):
 
     pythonexe = get_pythonexe_path()
     for dep in get_contrib_pysite_deps():
-        subprocess.call(
+        subprocess.check_call(
             [
                 pythonexe,
                 "-m",
                 "pip",
                 "install",
-                "--no-cache-dir",
+                # "--no-cache-dir",
                 "--no-compile",
                 "-t",
                 target_dir,
@@ -146,11 +158,11 @@ def get_contrib_pysite_deps():
     sys_type = util.get_systype()
     py_version = "%d%d" % (sys.version_info.major, sys.version_info.minor)
 
-    twisted_version = "19.7.0"
+    twisted_version = "19.10.0" if PY2 else "20.3.0"
     result = [
         "twisted == %s" % twisted_version,
-        "autobahn == 19.10.1",
-        "json-rpc == 1.12.1",
+        "autobahn == 20.4.3",
+        "json-rpc == 1.13.0",
     ]
 
     # twisted[tls], see setup.py for %twisted_version%
@@ -159,12 +171,12 @@ def get_contrib_pysite_deps():
     )
 
     # zeroconf
-    if sys.version_info.major < 3:
+    if PY2:
         result.append(
             "https://github.com/ivankravets/python-zeroconf/" "archive/pio-py27.zip"
         )
     else:
-        result.append("zeroconf == 0.23.0")
+        result.append("zeroconf == 0.26.0")
 
     if "windows" in sys_type:
         result.append("pypiwin32 == 223")