+++ tests/commands/test_account.py
@@ -20,6 +20,14 @@ import pytest
 
 from platformio.commands.account.command import cli as cmd_account
 
+pytestmark = pytest.mark.skipif(
+    not (
+        os.environ.get("PLATFORMIO_TEST_ACCOUNT_LOGIN")
+        and os.environ.get("PLATFORMIO_TEST_ACCOUNT_PASSWORD")
+    ),
+    reason="requires PLATFORMIO_TEST_ACCOUNT_LOGIN, PLATFORMIO_TEST_ACCOUNT_PASSWORD environ variables",
+)
+
 
 @pytest.fixture(scope="session")
 def credentials():
@@ -93,7 +101,7 @@ def test_account_login(clirunner, credentials, validate_cliresult, isolated_pio_
         )
         assert result.exit_code > 0
         assert result.exception
-        assert "You are already authenticated with" in str(result.exception)
+        assert "You are already authorized with" in str(result.exception)
     finally:
         clirunner.invoke(cmd_account, ["logout"])
 
@@ -113,7 +121,7 @@ def test_account_logout(clirunner, credentials, validate_cliresult, isolated_pio
         result = clirunner.invoke(cmd_account, ["logout"])
         assert result.exit_code > 0
         assert result.exception
-        assert "You are not authenticated! Please login to PIO Account" in str(
+        assert "You are not authorized! Please log in to PIO Account" in str(
             result.exception
         )
     finally:
@@ -192,7 +200,7 @@ def test_account_password_change(
         )
         assert result.exit_code > 0
         assert result.exception
-        assert "You are not authenticated! Please login to PIO Account" in str(
+        assert "You are not authorized! Please log in to PIO Account" in str(
             result.exception
         )
 
@@ -250,7 +258,7 @@ def test_account_token_with_invalid_password(
         )
         assert result.exit_code > 0
         assert result.exception
-        assert "You are not authenticated! Please login to PIO Account" in str(
+        assert "You are not authorized! Please log in to PIO Account" in str(
             result.exception
         )
 
@@ -302,7 +310,7 @@ def test_account_token(clirunner, credentials, validate_cliresult, isolated_pio_
         )
         assert result.exit_code > 0
         assert result.exception
-        assert "You are not authenticated! Please login to PIO Account" in str(
+        assert "You are not authorized! Please log in to PIO Account" in str(
             result.exception
         )
 
@@ -371,7 +379,7 @@ def test_account_summary(clirunner, credentials, validate_cliresult, isolated_pi
         result = clirunner.invoke(cmd_account, ["show"],)
         assert result.exit_code > 0
         assert result.exception
-        assert "You are not authenticated! Please login to PIO Account" in str(
+        assert "You are not authorized! Please log in to PIO Account" in str(
             result.exception
         )
 
@@ -380,6 +388,16 @@ def test_account_summary(clirunner, credentials, validate_cliresult, isolated_pi
             ["login", "-u", credentials["login"], "-p", credentials["password"]],
         )
         validate_cliresult(result)
+
+        result = clirunner.invoke(cmd_account, ["show", "--json-output", "--offline"])
+        validate_cliresult(result)
+        json_result = json.loads(result.output.strip())
+        assert not json_result.get("user_id")
+        assert json_result.get("profile")
+        assert json_result.get("profile").get("username")
+        assert json_result.get("profile").get("email")
+        assert not json_result.get("packages")
+        assert not json_result.get("subscriptions")
 
         result = clirunner.invoke(cmd_account, ["show"])
         validate_cliresult(result)
@@ -407,12 +425,19 @@ def test_account_summary(clirunner, credentials, validate_cliresult, isolated_pi
         result = clirunner.invoke(cmd_account, ["show", "--json-output", "--offline"])
         validate_cliresult(result)
         json_result = json.loads(result.output.strip())
-        assert not json_result.get("user_id")
+        assert json_result.get("user_id")
         assert json_result.get("profile")
         assert json_result.get("profile").get("username")
         assert json_result.get("profile").get("email")
-        assert not json_result.get("packages")
-        assert not json_result.get("subscriptions")
+        assert credentials["login"] == json_result.get("profile").get(
+            "username"
+        ) or credentials["login"] == json_result.get("profile").get("email")
+        assert json_result.get("profile").get("firstname")
+        assert json_result.get("profile").get("lastname")
+        assert json_result.get("packages")
+        assert json_result.get("packages")[0].get("name")
+        assert json_result.get("packages")[0].get("path")
+        assert json_result.get("subscriptions") is not None
     finally:
         clirunner.invoke(cmd_account, ["logout"])
 
@@ -427,7 +452,7 @@ def test_account_profile_update_with_invalid_password(
         )
         assert result.exit_code > 0
         assert result.exception
-        assert "You are not authenticated! Please login to PIO Account" in str(
+        assert "You are not authorized! Please log in to PIO Account" in str(
             result.exception
         )
 
@@ -460,7 +485,7 @@ def test_account_profile_update_only_firstname_and_lastname(
         )
         assert result.exit_code > 0
         assert result.exception
-        assert "You are not authenticated! Please login to PIO Account" in str(
+        assert "You are not authorized! Please log in to PIO Account" in str(
             result.exception
         )
 
@@ -508,7 +533,7 @@ def test_account_profile_update(
         )
         assert result.exit_code > 0
         assert result.exception
-        assert "You are not authenticated! Please login to PIO Account" in str(
+        assert "You are not authorized! Please log in to PIO Account" in str(
             result.exception
         )
 
@@ -549,7 +574,7 @@ def test_account_profile_update(
         result = clirunner.invoke(cmd_account, ["show"],)
         assert result.exit_code > 0
         assert result.exception
-        assert "You are not authenticated! Please login to PIO Account" in str(
+        assert "You are not authorized! Please log in to PIO Account" in str(
             result.exception
         )